Create common virtualization editor component - fix `useVirtualization` so that polling does not call read callback if there have not been any changes in the virualization - created `VirtualizationEditorPage` that contains common code used by all editor pages - removed virtualization name from callback properties in `ViewHeaderBreadcrumb` as it is not needed - modified `VirtualizationViewsPage` and `VirtualizationSqlClientPage` to use `VirtualizationEditorPage` and add their custom UI to it - added polling to `useViewDefinitionDescriptors` to get view descriptor updates
