+++ app/ui-react/packages/api/src/useVirtualization.tsx
@@ -1,12 +1,20 @@
 import { Virtualization } from '@syndesis/models';
+import * as React from 'react';
+import isEqual from 'react-fast-compare';
 import { useApiResource } from './useApiResource';
 import { usePolling } from './usePolling';
 
+/**
+ * @property {Virtualization} model updates only when the value of `resource` changes
+ * @property {Virtualization} resource updates each time fetch is called
+ * @param virtualizationName the name of the virtualization being requested
+ * @param disableUpdates `true` if polling should be enabled (defaults to `false`)
+ */
 export const useVirtualization = (
   virtualizationName: string,
   disableUpdates: boolean = false
 ) => {
-  const { read, ...rest } = useApiResource<Virtualization>({
+  const { read, resource, ...rest } = useApiResource<Virtualization>({
     defaultValue: {
       description: '',
       empty: true,
@@ -20,12 +28,27 @@ export const useVirtualization = (
     useDvApiUrl: true,
   });
 
-  if (!disableUpdates) {
-    usePolling({ callback: read, delay: 5000 });
-  }
+  /**
+   * Update `model` only if `resource` changes.
+   */
+  const [model, setModel] = React.useState(resource);
+  React.useEffect(() => {
+    if (!isEqual(resource, model)) {
+      setModel(resource);
+    }
+  }, [resource, model, setModel]);
+
+  // only poll if updates are turned on
+  const poller = React.useCallback(() => {
+    // tslint:disable-next-line: no-unused-expression
+     disableUpdates ? () => void 0 : read();
+  }, []);
+  usePolling({ callback: poller, delay: 5000 });
 
   return {
     ...rest,
+    model, // updates only when the resource changes
     read,
+    resource, // updates each time polling runs
   };
 };
