+++ app/ui-react/syndesis/src/modules/data/pages/VirtualizationEditorPage.css
@@ -1,10 +1,10 @@
-.virtualization-sql-client-page .inline-text-editwidget textarea {
+.virtualization-editor-page .inline-text-editwidget textarea {
   height: 8.5em;
   font-size: 1.0em;
   line-height: 1.25em;
 }
 
-.virtualization-sql-client-page .virtualization-details-header .inline-text-readwidget {
+.virtualization-editor-page .virtualization-details-header .inline-text-readwidget {
   display: inline-block;
   line-height: 1.25em;
   max-height: 6.25em;
