+++ app/ui-react/syndesis/src/modules/data/pages/index.ts
@@ -1,4 +1,5 @@
 export * from './VirtualizationCreatePage';
+export * from './VirtualizationEditorPage';
 export * from './VirtualizationImportPage';
 export * from './VirtualizationsPage';
 export * from './VirtualizationSqlClientPage';
